requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
Pillow
piexif
pillow-heif